feat: Add Monolog for logging
This commit introduces the Monolog library as a dependency for enhanced logging capabilities within the application.

The following changes were made to `composer.json`:

- Added `"monolog/monolog": "^3.9"` to the `require` section.

This new dependency will allow for more structured and flexible logging, which is crucial for debugging and monitoring the application. To enable the use of Monolog, a subsequent `composer install` or `composer update` command will need to be executed.